refactor: move VideoPoster component
