+++ web/components/video/VideoPoster/VideoPoster.module.scss
